safari-600.7-branch/LayoutTests/platform/ios-sim/fast/dom/custom/lifecycle-ready-innerHTML-expected.txt

CONSOLE MESSAGE: line 16: TypeError: undefined is not a function (evaluating 'document.register("x-foo", { prototype: Object.create(HTMLElement.prototype, { readyCallback: { value: trackingReadyCallback } }) })')
This test ensures that the lifecycle callback of a parser-made element is visible in following script block.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL window.callbacksCalled should be a,b,c. Was .
FAIL window.callbacksCalled should be a,b,c. Was .
FAIL successfullyParsed should be true (of type boolean). Was undefined (of type undefined).

TEST COMPLETE
